mpc_orb/requirements.txt
attrs==21.2.0
certifi==2021.5.30
jsonschema==3.2.0
numpy==1.19.5
pyrsistent==0.18.0
six==1.16.0
